fix: Forgot send get headers when received UnknownParent compact block
