+++ sync/src/relayer/compact_block_process.rs
@@ -40,19 +40,35 @@ impl<'a, CS: ChainStore> CompactBlockProcess<'a, CS> {
     pub fn execute(self) -> Result<(), FailureError> {
         let compact_block: CompactBlock = (*self.message).try_into()?;
         let block_hash = compact_block.header.hash();
+        if let Some(header_view) = self.relayer.shared.get_header_view(&block_hash) {
+            let best_known_header = self.relayer.shared.best_known_header();
+            if header_view.total_difficulty() <= best_known_header.total_difficulty() {
+                debug!(
+                    target: "relay",
+                    "Received a compact block({}), total difficulty {} <= {}",
+                    block_hash,
+                    header_view.total_difficulty(),
+                    best_known_header.total_difficulty(),
+                );
+                return Ok(());
+            }
+        }
+
         let mut missing_indexes: Vec<usize> = Vec::new();
         {
             let mut pending_compact_blocks = self.relayer.state.pending_compact_blocks.lock();
             if pending_compact_blocks.get(&block_hash).is_none()
-                && self.relayer.get_block(&block_hash).is_none()
+                && self.relayer.shared.get_block(&block_hash).is_none()
             {
-                let resolver =
-                    HeaderResolverWrapper::new(&compact_block.header, self.relayer.shared.clone());
+                let resolver = HeaderResolverWrapper::new(
+                    &compact_block.header,
+                    self.relayer.shared.shared().clone(),
+                );
                 let header_verifier = HeaderVerifier::new(
                     CompactBlockMedianTimeView {
                         header: &compact_block.header,
                         pending_compact_blocks: &pending_compact_blocks,
-                        shared: &self.relayer.shared,
+                        shared: self.relayer.shared.shared(),
                     },
                     Arc::clone(&self.relayer.shared.consensus().pow_engine()),
                 );
@@ -83,7 +99,7 @@ impl<'a, CS: ChainStore> CompactBlockProcess<'a, CS> {
                     }
                     Err(VerifyError::UnknownParent(hash)) => {
                         debug!(target: "relay", "UnknownParent: {}, send_getheaders_to_peer({})", hash, self.peer);
-                        self.relayer.send_getheaders_to_peer(
+                        self.relayer.shared.send_getheaders_to_peer(
                             self.nc,
                             self.peer,
                             self.relayer.shared.chain_state().lock().tip_header(),
