+++ sync/src/relayer/get_block_transactions_process.rs
@@ -35,7 +35,7 @@ impl<'a, CS: ChainStore> GetBlockTransactionsProcess<'a, CS> {
 
         let indexes = cast!(self.message.indexes())?;
 
-        if let Some(block) = self.relayer.get_block(&block_hash) {
+        if let Some(block) = self.relayer.shared.get_block(&block_hash) {
             let transactions = indexes
                 .safe_slice()
                 .iter()
