+++ sync/src/synchronizer/get_headers_process.rs
@@ -36,7 +36,7 @@ where
     }
 
     pub fn execute(self) -> Result<(), FailureError> {
-        if self.synchronizer.is_initial_block_download() {
+        if self.synchronizer.shared.is_initial_block_download() {
             info!(target: "sync", "Ignoring getheaders from peer={} because node is in initial block download", self.peer);
             return Ok(());
         }
@@ -56,13 +56,20 @@ where
 
         if let Some(block_number) = self
             .synchronizer
+            .shared
             .locate_latest_common_block(&hash_stop, &block_locator_hashes[..])
         {
-            debug!(target: "sync", "\n\nheaders latest_common={} tip={} begin\n\n", block_number, {self.synchronizer.tip_header().number()});
+            debug!(
+                target: "sync",
+                "\n\nheaders latest_common={} tip={} begin\n\n",
+                block_number,
+                self.synchronizer.shared.tip_header().number(),
+            );
 
             self.synchronizer.peers.getheaders_received(self.peer);
             let headers: Vec<Header> = self
                 .synchronizer
+                .shared
                 .get_locator_response(block_number, &hash_stop);
             // response headers
