+++ sync/src/synchronizer/mod.rs
@@ -12,29 +12,25 @@ use self::get_blocks_process::GetBlocksProcess;
 use self::get_headers_process::GetHeadersProcess;
 use self::headers_process::HeadersProcess;
 use crate::config::Config;
-use crate::types::{HeaderView, Peers};
+use crate::types::{HeaderView, Peers, SyncSharedState};
 use crate::{
     BAD_MESSAGE_BAN_TIME, CHAIN_SYNC_TIMEOUT, EVICTION_HEADERS_RESPONSE_TIME,
-    HEADERS_DOWNLOAD_TIMEOUT_BASE, HEADERS_DOWNLOAD_TIMEOUT_PER_HEADER, MAX_HEADERS_LEN,
-    MAX_OUTBOUND_PEERS_TO_PROTECT_FROM_DISCONNECT, MAX_TIP_AGE, POW_SPACE,
+    HEADERS_DOWNLOAD_TIMEOUT_BASE, HEADERS_DOWNLOAD_TIMEOUT_PER_HEADER,
+    MAX_OUTBOUND_PEERS_TO_PROTECT_FROM_DISCONNECT, POW_SPACE,
 };
 use bitflags::bitflags;
 use ckb_chain::chain::ChainController;
-use ckb_chain_spec::consensus::Consensus;
 use ckb_core::block::Block;
-use ckb_core::header::{BlockNumber, Header};
+use ckb_core::header::Header;
 use ckb_network::{CKBProtocolContext, CKBProtocolHandler, PeerIndex};
 use ckb_protocol::{cast, get_root, SyncMessage, SyncPayload};
-use ckb_shared::shared::Shared;
 use ckb_shared::store::ChainStore;
-use ckb_traits::ChainProvider;
-use ckb_util::{try_option, Mutex, RwLock};
+use ckb_util::Mutex;
 use failure::Error as FailureError;
 use faketime::unix_time_as_millis;
 use flatbuffers::FlatBufferBuilder;
 use log::{debug, info, trace};
 use numext_fixed_hash::H256;
-use std::cmp;
 use std::collections::HashMap;
 use std::sync::atomic::AtomicUsize;
 use std::sync::atomic::Ordering;
@@ -66,14 +62,11 @@ bitflags! {
 }
 
 pub type BlockStatusMap = Arc<Mutex<HashMap<H256, BlockStatus>>>;
-pub type BlockHeaderMap = Arc<RwLock<HashMap<H256, HeaderView>>>;
 
 pub struct Synchronizer<CS: ChainStore> {
     chain: ChainController,
-    shared: Shared<CS>,
+    shared: Arc<SyncSharedState<CS>>,
     pub status_map: BlockStatusMap,
-    pub header_map: BlockHeaderMap,
-    pub best_known_header: Arc<RwLock<HeaderView>>,
     pub n_sync: Arc<AtomicUsize>,
     pub peers: Arc<Peers>,
     pub config: Arc<Config>,
@@ -86,10 +79,8 @@ impl<CS: ChainStore> ::std::clone::Clone for Synchronizer<CS> {
     fn clone(&self) -> Self {
         Synchronizer {
             chain: self.chain.clone(),
-            shared: self.shared.clone(),
+            shared: Arc::clone(&self.shared),
             status_map: Arc::clone(&self.status_map),
-            header_map: Arc::clone(&self.header_map),
-            best_known_header: Arc::clone(&self.best_known_header),
             n_sync: Arc::clone(&self.n_sync),
             peers: Arc::clone(&self.peers),
             config: Arc::clone(&self.config),
@@ -100,30 +91,19 @@ impl<CS: ChainStore> ::std::clone::Clone for Synchronizer<CS> {
 }
 
 impl<CS: ChainStore> Synchronizer<CS> {
-    pub fn new(chain: ChainController, shared: Shared<CS>, config: Config) -> Synchronizer<CS> {
-        let (total_difficulty, header, total_uncles_count) = {
-            let chain_state = shared.chain_state().lock();
-            let block_ext = shared
-                .block_ext(&chain_state.tip_hash())
-                .expect("tip block_ext must exist");
-            (
-                chain_state.total_difficulty().clone(),
-                chain_state.tip_header().clone(),
-                block_ext.total_uncles_count,
-            )
-        };
-        let best_known_header = HeaderView::new(header, total_difficulty, total_uncles_count);
+    pub fn new(
+        chain: ChainController,
+        shared: Arc<SyncSharedState<CS>>,
+        config: Config,
+    ) -> Synchronizer<CS> {
         let orphan_block_limit = config.orphan_block_limit;
-
         Synchronizer {
             config: Arc::new(config),
             chain,
             shared,
             peers: Arc::new(Peers::default()),
             orphan_block_pool: Arc::new(OrphanBlockPool::with_capacity(orphan_block_limit)),
-            best_known_header: Arc::new(RwLock::new(best_known_header)),
             status_map: Arc::new(Mutex::new(HashMap::new())),
-            header_map: Arc::new(RwLock::new(HashMap::new())),
             n_sync: Arc::new(AtomicUsize::new(0)),
             outbound_peers_with_protect: Arc::new(AtomicUsize::new(0)),
         }
@@ -190,16 +170,6 @@ impl<CS: ChainStore> Synchronizer<CS> {
         self.status_map.lock().insert(hash, status);
     }
 
-    pub fn best_known_header(&self) -> HeaderView {
-        self.best_known_header.read().clone()
-    }
-
-    pub fn is_initial_block_download(&self) -> bool {
-        unix_time_as_millis()
-            .saturating_sub(self.shared.chain_state().lock().tip_header().timestamp())
-            > MAX_TIP_AGE
-    }
-
     pub fn predict_headers_sync_time(&self, header: &Header) -> u64 {
         let now = unix_time_as_millis();
         now + HEADERS_DOWNLOAD_TIMEOUT_BASE
@@ -215,156 +185,13 @@ impl<CS: ChainStore> Synchronizer<CS> {
             .or_insert_with(|| BlockStatus::BLOCK_HAVE_MASK);
     }
 
-    pub fn tip_header(&self) -> Header {
-        self.shared.chain_state().lock().tip_header().clone()
-    }
-
-    pub fn get_locator(&self, start: &Header) -> Vec<H256> {
-        let mut step = 1;
-        let mut locator = Vec::with_capacity(32);
-        let mut index = start.number();
-        let base = start.hash();
-        loop {
-            let header = self
-                .get_ancestor(&base, index)
-                .expect("index calculated in get_locator");
-            locator.push(header.hash().clone());
-
-            if locator.len() >= 10 {
-                step <<= 1;
-            }
-
-            if index < step {
-                // always include genesis hash
-                if index != 0 {
-                    locator.push(self.shared.genesis_hash().clone());
-                }
-                break;
-            }
-            index -= step;
-        }
-        locator
-    }
-
-    pub fn locate_latest_common_block(
-        &self,
-        _hash_stop: &H256,
-        locator: &[H256],
-    ) -> Option<BlockNumber> {
-        if locator.is_empty() {
-            return None;
-        }
-
-        if locator.last().expect("empty checked") != self.shared.genesis_hash() {
-            return None;
-        }
-
-        // iterator are lazy
-        let (index, latest_common) = locator
-            .iter()
-            .enumerate()
-            .map(|(index, hash)| (index, self.shared.block_number(hash)))
-            .find(|(_index, number)| number.is_some())
-            .expect("locator last checked");
-
-        if index == 0 || latest_common == Some(0) {
-            return latest_common;
-        }
-
-        if let Some(header) = locator
-            .get(index - 1)
-            .and_then(|hash| self.shared.block_header(hash))
-        {
-            let mut block_hash = header.parent_hash().clone();
-            loop {
-                let block_header = match self.shared.block_header(&block_hash) {
-                    None => break latest_common,
-                    Some(block_header) => block_header,
-                };
-
-                if let Some(block_number) = self.shared.block_number(&block_hash) {
-                    return Some(block_number);
-                }
-
-                block_hash = block_header.parent_hash().clone();
-            }
-        } else {
-            latest_common
-        }
-    }
-
-    pub fn get_header_view(&self, hash: &H256) -> Option<HeaderView> {
-        self.header_map.read().get(hash).cloned().or_else(|| {
-            self.shared.block_header(hash).and_then(|header| {
-                self.shared.block_ext(&hash).map(|block_ext| {
-                    HeaderView::new(
-                        header,
-                        block_ext.total_difficulty,
-                        block_ext.total_uncles_count,
-                    )
-                })
-            })
-        })
-    }
-
-    pub fn consensus(&self) -> &Consensus {
-        self.shared.consensus()
-    }
-
-    pub fn get_header(&self, hash: &H256) -> Option<Header> {
-        self.header_map
-            .read()
-            .get(hash)
-            .map(HeaderView::inner)
-            .cloned()
-            .or_else(|| self.shared.block_header(hash))
-    }
-
-    pub fn get_block(&self, hash: &H256) -> Option<Block> {
-        self.shared.block(hash)
-    }
-
-    pub fn get_ancestor(&self, base: &H256, number: BlockNumber) -> Option<Header> {
-        if let Some(header) = self.get_header(base) {
-            let mut n_number = header.number();
-            let mut index_walk = header;
-            if number > n_number {
-                return None;
-            }
-
-            while n_number > number {
-                if let Some(header) = self.get_header(&index_walk.parent_hash()) {
-                    index_walk = header;
-                    n_number -= 1;
-                } else {
-                    return None;
-                }
-            }
-            return Some(index_walk);
-        }
-        None
-    }
-
-    pub fn get_locator_response(&self, block_number: BlockNumber, hash_stop: &H256) -> Vec<Header> {
-        let tip_number = self.tip_header().number();
-        let max_height = cmp::min(
-            block_number + 1 + MAX_HEADERS_LEN as BlockNumber,
-            tip_number + 1,
-        );
-        (block_number + 1..max_height)
-            .filter_map(|block_number| self.shared.block_hash(block_number))
-            .take_while(|block_hash| block_hash != hash_stop)
-            .filter_map(|block_hash| self.shared.block_header(&block_hash))
-            .collect()
-    }
-
     pub fn insert_header_view(&self, header: &Header, peer: PeerIndex) {
-        if let Some(parent_view) = self.get_header_view(&header.parent_hash()) {
+        if let Some(parent_view) = self.shared.get_header_view(&header.parent_hash()) {
             let total_difficulty = parent_view.total_difficulty() + header.difficulty();
             let total_uncles_count =
                 parent_view.total_uncles_count() + u64::from(header.uncles_count());
             let header_view = {
-                let mut best_known_header = self.best_known_header.write();
+                let best_known_header = self.shared.best_known_header();
                 let header_view =
                     HeaderView::new(header.clone(), total_difficulty.clone(), total_uncles_count);
 
@@ -372,42 +199,15 @@ impl<CS: ChainStore> Synchronizer<CS> {
                     || (&total_difficulty == best_known_header.total_difficulty()
                         && header.hash() < best_known_header.hash())
                 {
-                    *best_known_header = header_view.clone();
+                    self.shared.set_best_known_header(header_view.clone());
                 }
                 header_view
             };
 
             self.peers.new_header_received(peer, &header_view);
-
-            let mut header_map = self.header_map.write();
-            header_map.insert(header.hash().clone(), header_view);
+            self.shared
+                .insert_header(header.hash().clone(), header_view);
         }
-    }
-
-    // If the peer reorganized, our previous last_common_header may not be an ancestor
-    // of its current best_known_header. Go back enough to fix that.
-    pub fn last_common_ancestor(
-        &self,
-        last_common_header: &Header,
-        best_known_header: &Header,
-    ) -> Option<Header> {
-        debug_assert!(best_known_header.number() >= last_common_header.number());
-
-        let mut m_right =
-            try_option!(self.get_ancestor(&best_known_header.hash(), last_common_header.number()));
-
-        if &m_right == last_common_header {
-            return Some(m_right);
-        }
-
-        let mut m_left = try_option!(self.get_header(&last_common_header.hash()));
-        debug_assert!(m_right.number() == m_left.number());
-
-        while m_left != m_right {
-            m_left = try_option!(self.get_ancestor(&m_left.hash(), m_left.number() - 1));
-            m_right = try_option!(self.get_ancestor(&m_right.hash(), m_right.number() - 1));
-        }
-        Some(m_left)
     }
 
     //TODO: process block which we don't request
@@ -494,7 +294,7 @@ impl<CS: ChainStore> Synchronizer<CS> {
     }
 
     fn on_connected(&self, nc: &CKBProtocolContext, peer: PeerIndex) {
-        let tip = self.tip_header();
+        let tip = self.shared.tip_header();
         let predicted_headers_sync_time = self.predict_headers_sync_time(&tip);
 
         let is_outbound = nc
@@ -514,20 +314,6 @@ impl<CS: ChainStore> Synchronizer<CS> {
             .on_connected(peer, predicted_headers_sync_time, protect_outbound);
     }
 
-    pub fn send_getheaders_to_peer(
-        &self,
-        nc: &CKBProtocolContext,
-        peer: PeerIndex,
-        header: &Header,
-    ) {
-        debug!(target: "sync", "send_getheaders_to_peer peer={}, hash={}", peer, header.hash());
-        let locator_hash = self.get_locator(header);
-        let fbb = &mut FlatBufferBuilder::new();
-        let message = SyncMessage::build_get_headers(fbb, &locator_hash);
-        fbb.finish(message, None);
-        nc.send_message_to(peer, fbb.finished_data().to_vec());
-    }
-
     //   - If at timeout their best known block now has more work than our tip
     //     when the timeout was set, then either reset the timeout or clear it
     //     (after comparing against our current tip's work)
@@ -539,7 +325,7 @@ impl<CS: ChainStore> Synchronizer<CS> {
     pub fn eviction(&self, nc: &CKBProtocolContext) {
         let mut peer_state = self.peers.state.write();
         let best_known_headers = self.peers.best_known_headers.read();
-        let is_initial_block_download = self.is_initial_block_download();
+        let is_initial_block_download = self.shared.is_initial_block_download();
         let mut eviction = Vec::new();
         for (peer, state) in peer_state.iter_mut() {
             let now = unix_time_as_millis();
@@ -595,7 +381,7 @@ impl<CS: ChainStore> Synchronizer<CS> {
                         } else {
                             state.chain_sync.sent_getheaders = true;
                             state.chain_sync.timeout = now + EVICTION_HEADERS_RESPONSE_TIME;
-                            self.send_getheaders_to_peer(
+                            self.shared.send_getheaders_to_peer(
                                 nc,
                                 *peer,
                                 &state
@@ -634,7 +420,7 @@ impl<CS: ChainStore> Synchronizer<CS> {
                     chain_state.total_difficulty().clone(),
                 )
             };
-            let best_known = self.best_known_header();
+            let best_known = self.shared.best_known_header();
             if total_difficulty > *best_known.total_difficulty()
                 || (&total_difficulty == best_known.total_difficulty()
                     && header.hash() < best_known.hash())
@@ -646,7 +432,7 @@ impl<CS: ChainStore> Synchronizer<CS> {
         };
         for peer in peers {
             // Only sync with 1 peer if we're in IBD
-            if self.is_initial_block_download() && self.n_sync.load(Ordering::Acquire) != 0 {
+            if self.shared.is_initial_block_download() && self.n_sync.load(Ordering::Acquire) != 0 {
                 break;
             }
             {
@@ -660,7 +446,7 @@ impl<CS: ChainStore> Synchronizer<CS> {
             }
 
             debug!(target: "sync", "start sync peer={}", peer);
-            self.send_getheaders_to_peer(nc, peer, &tip);
+            self.shared.send_getheaders_to_peer(nc, peer, &tip);
         }
     }
 
@@ -766,9 +552,11 @@ mod tests {
     use self::block_process::BlockProcess;
     use self::headers_process::HeadersProcess;
     use super::*;
+    use crate::{SyncSharedState, MAX_TIP_AGE};
     use ckb_chain::chain::ChainBuilder;
     use ckb_chain_spec::consensus::Consensus;
     use ckb_core::block::BlockBuilder;
+    use ckb_core::header::BlockNumber;
     use ckb_core::header::{Header, HeaderBuilder};
     use ckb_core::script::Script;
     use ckb_core::transaction::{CellInput, CellOutput, Transaction, TransactionBuilder};
@@ -780,8 +568,10 @@ mod tests {
     };
     use ckb_notify::{NotifyController, NotifyService};
     use ckb_protocol::{Block as FbsBlock, Headers as FbsHeaders};
+    use ckb_shared::shared::Shared;
     use ckb_shared::shared::SharedBuilder;
     use ckb_shared::store::{ChainKVStore, ChainStore};
+    use ckb_traits::chain_provider::ChainProvider;
     use ckb_util::Mutex;
     #[cfg(not(disable_faketime))]
     use faketime;
@@ -818,6 +608,7 @@ mod tests {
         chain_controller: ChainController,
         shared: Shared<CS>,
     ) -> Synchronizer<CS> {
+        let shared = Arc::new(SyncSharedState::new(shared));
         Synchronizer::new(chain_controller, shared, Config::default())
     }
 
@@ -889,7 +680,9 @@ mod tests {
 
         let synchronizer = gen_synchronizer(chain_controller.clone(), shared.clone());
 
-        let locator = synchronizer.get_locator(shared.chain_state().lock().tip_header());
+        let locator = synchronizer
+            .shared
+            .get_locator(shared.chain_state().lock().tip_header());
 
         let mut expect = Vec::new();
 
@@ -921,9 +714,13 @@ mod tests {
 
         let synchronizer2 = gen_synchronizer(chain_controller2.clone(), shared2.clone());
 
-        let locator1 = synchronizer1.get_locator(shared1.chain_state().lock().tip_header());
+        let locator1 = synchronizer1
+            .shared
+            .get_locator(shared1.chain_state().lock().tip_header());
 
-        let latest_common = synchronizer2.locate_latest_common_block(&H256::zero(), &locator1[..]);
+        let latest_common = synchronizer2
+            .shared
+            .locate_latest_common_block(&H256::zero(), &locator1[..]);
 
         assert_eq!(latest_common, Some(0));
 
@@ -936,7 +733,9 @@ mod tests {
 
         let synchronizer3 = gen_synchronizer(chain_controller3.clone(), shared3.clone());
 
-        let latest_common3 = synchronizer3.locate_latest_common_block(&H256::zero(), &locator1[..]);
+        let latest_common3 = synchronizer3
+            .shared
+            .locate_latest_common_block(&H256::zero(), &locator1[..]);
         assert_eq!(latest_common3, Some(192));
     }
 
@@ -977,9 +776,12 @@ mod tests {
 
         let synchronizer1 = gen_synchronizer(chain_controller1.clone(), shared1.clone());
         let synchronizer2 = gen_synchronizer(chain_controller2.clone(), shared2.clone());
-        let locator1 = synchronizer1.get_locator(shared1.chain_state().lock().tip_header());
+        let locator1 = synchronizer1
+            .shared
+            .get_locator(shared1.chain_state().lock().tip_header());
 
         let latest_common = synchronizer2
+            .shared
             .locate_latest_common_block(&H256::zero(), &locator1[..])
             .unwrap();
 
@@ -1006,9 +808,15 @@ mod tests {
 
         let synchronizer = gen_synchronizer(chain_controller.clone(), shared.clone());
 
-        let header = synchronizer.get_ancestor(&shared.chain_state().lock().tip_hash(), 100);
-        let tip = synchronizer.get_ancestor(&shared.chain_state().lock().tip_hash(), 199);
-        let noop = synchronizer.get_ancestor(&shared.chain_state().lock().tip_hash(), 200);
+        let header = synchronizer
+            .shared
+            .get_ancestor(&shared.chain_state().lock().tip_hash(), 100);
+        let tip = synchronizer
+            .shared
+            .get_ancestor(&shared.chain_state().lock().tip_hash(), 199);
+        let noop = synchronizer
+            .shared
+            .get_ancestor(&shared.chain_state().lock().tip_hash(), 200);
         assert!(tip.is_some());
         assert!(header.is_some());
         assert!(noop.is_none());
@@ -1078,7 +886,7 @@ mod tests {
 
         let synchronizer = gen_synchronizer(chain_controller.clone(), shared.clone());
 
-        let headers = synchronizer.get_locator_response(180, &H256::zero());
+        let headers = synchronizer.shared.get_locator_response(180, &H256::zero());
 
         assert_eq!(headers.first().unwrap(), blocks[180].header());
         assert_eq!(headers.last().unwrap(), blocks[199].header());
@@ -1119,6 +927,7 @@ mod tests {
         fn set_notify(&self, _interval: Duration, _token: u64) {
             unimplemented!();
         }
+        fn send_message(&self, _proto_id: ProtocolId, _peer_index: PeerIndex, _data: Vec<u8>) {}
         fn send_message_to(&self, _peer_index: PeerIndex, _data: Vec<u8>) {}
         fn filter_broadcast(&self, _target: TargetSession, _data: Vec<u8>) {
             unimplemented!();
@@ -1169,7 +978,9 @@ mod tests {
 
         let synchronizer1 = gen_synchronizer(chain_controller1.clone(), shared1.clone());
 
-        let locator1 = synchronizer1.get_locator(&shared1.chain_state().lock().tip_header());
+        let locator1 = synchronizer1
+            .shared
+            .get_locator(&shared1.chain_state().lock().tip_header());
 
         for i in 1..=num {
             let j = if i > 192 { i + 1 } else { i };
@@ -1177,10 +988,14 @@ mod tests {
         }
 
         let synchronizer2 = gen_synchronizer(chain_controller2.clone(), shared2.clone());
-        let latest_common = synchronizer2.locate_latest_common_block(&H256::zero(), &locator1[..]);
+        let latest_common = synchronizer2
+            .shared
+            .locate_latest_common_block(&H256::zero(), &locator1[..]);
         assert_eq!(latest_common, Some(192));
 
-        let headers = synchronizer2.get_locator_response(192, &H256::zero());
+        let headers = synchronizer2
+            .shared
+            .get_locator_response(192, &H256::zero());
 
         assert_eq!(
             headers.first().unwrap().hash(),
@@ -1271,7 +1086,7 @@ mod tests {
 
         let network_context = mock_network_context(5);
         faketime::write_millis(&faketime_file, MAX_TIP_AGE * 2).expect("write millis");
-        assert!(synchronizer.is_initial_block_download());
+        assert!(synchronizer.shared.is_initial_block_download());
         let peers = synchronizer.peers();
         // protect should not effect headers_timeout
         peers.on_connected(0.into(), 0, true);
